src/main/java/com/ecommerce/product/infrastructure/web/exception/ExceptionHandlerAdvice.java
package com.ecommerce.product.infrastructure.web.exception;

import com.ecommerce.product.domain.exception.DomainException;
import com.ecommerce.product.infrastructure.web.dto.ErrorResponseDTO;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.ControllerAdvice;
import org.springframework.web.bind.annotation.ExceptionHandler;

@ControllerAdvice
public class ExceptionHandlerAdvice {

    @ExceptionHandler(DomainException.class)
    public ResponseEntity<ErrorResponseDTO> handleDomainException(Exception ex) {
        ErrorResponseDTO response = new ErrorResponseDTO();
        response.setError(HttpStatus.INTERNAL_SERVER_ERROR.getReasonPhrase());
        response.setMessage(ex.getMessage());
        return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(response);
    }
}